Changed: Use BigDecimal instead of Double

+++ 5. testing/ObtenerDiploma/src/main/java/com/meli/obtenerdiploma/model/StudentDTO.java
@@ -8,6 +8,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.Pattern;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -22,7 +23,7 @@ public class StudentDTO {
 
     private String message;
 
-    private Double averageScore;
+    private BigDecimal averageScore;
 
     @NotEmpty(message = "La lista no puede ser vacía.")
     private List<@Valid SubjectDTO> subjects;
